refactor: mv net/InverseSockets/SslLogger => net/SslLogger
